utils/flujo_completo/requirements.txt
boto3
pandas
matplotlib
Pillow
sklearn
numpy
pdf2image
fuzzywuzzy
unidecode
nltk
